move CutoffStrategy to corresponding file in core
