+++ trane/core/__init__.py
@@ -1,4 +1,6 @@
 from .prediction_problem_generator import *  # noqa
 from .prediction_problem import *  # noqa
+from .cutoff_strategy import CutoffStrategy # noqa
 from .labeler import *  # noqa
 from .prediction_problem_saver import *  # noqa
+
