+++ trane/utils/__init__.py
@@ -1,4 +1,3 @@
 from .table_meta import *  # noqa
-from .generate_cutoff_times import CutoffStrategy # noqa
 from .data_parser import *  # noqa
 # from .generate_nl_description import *  # noqa
